fix: zoomed in screen sizes could lose text
Issue
=====

When cypress was used in a browser that was zoomed in, or otherwise used
a screen size that conflicted with the xterm.js size calculation logic,
the terminal would be too small and some text could be lost.

This was caused by ignoring the client's terminal size when starting
a new TerminalApplication instance, and using a hardcoded default size
instead.

Solution
========

Use the client's terminal size when starting a new TerminalApplication
so that no text is lost.