feat(wip): allow access to neovim via a type safe socket connection
This change is a bit of work-in-progress effort. It does not actually
expose the client yet. I'll have to think how it could be done, since we
have to run the tests in the browser, but the client exists on the
backend.

Right now I want to roll this out to see if causes any stability issues.
It's lazy so it should basically do nothing right now.