feat!: automatically generate cypress support bindings on startup
BREAKING CHANGE: this changes the way that cypress support bindings are
generated. Previously, you had to manually create a file in your
cypress/support directory. Now, the file is automatically generated for
you when you run `tui start`.

To compensate for this change, you need to update your e2e.ts file to
import the autogenerated file instead of the old, custom one.

The benefit of this change is that you no longer need to maintain your
own types for all future tui-sandbox versions.