feat: can run a headless neovim ex-command before tests
Issue
=====

Sometimes when starting the test run, the neovim dependencies (plugins
and such) are not installed yet. This is done automatically when neovim
starts, but it can be slow. This can cause the tests to fail because
they depend on the plugins.

Solution
========

Add a new command to `tui neovim exec '<ex-command>'` that will run the
given ex-command in a headless neovim instance before starting the tests.

This can be used to initialize the plugins with
`tui neovim exec "Lazy! sync"`