feat!: make it more difficult to accidentally call unsupported cmds
Issue
=====

The `cy.runBlockingShellCommand`, `cy.runLuaCode`, and `cy.runExCommand`
commands can only be called when a Neovim instance is running. However,
it is knowledge that you might not have beforehand, making it confusing
for new users.

Solution
========

Return a scoped "NeovimContext" object from `cy.startNeovim` that
provides access to the commands that require a running Neovim instance.
This way, the commands can only be called when a Neovim instance is
available, making it more difficult to call them accidentally.

BREAKING CHANGE: When a Neovim instance is started, tui-sandbox used to
return the TestDirectory instance, which is the type-safe representation
of the test directory's paths. The TestDirectory can now be found in a
`.dir` property.

BREAKING CHANGE: The `cy.runBlockingShellCommand`, `cy.runLuaCode`, and
`cy.runExCommand` commands are now available on the NeovimContext
object. Alternatively, they can still be accessed through the following:
- `cy.nvim_runBlockingShellCommand`
- `cy.nvim_runLuaCode`
- `cy.nvim_runExCommand`