feat: add nvim_isRunning command to check if neovim is running (#382)
This is useful when tests are too flaky and need to be left out of CI,
but can still be run locally on the developers' machines to offer some
benefits.

In such a case, using afterEach hooks might not be possible, because
they will often rely on neovim having been started.