perf: use zod v4 beta for faster type checking
This has been a minor annoyance when working with large schemas, such as
schemas generated for MyTestDirectorySchema in big projects. Hopefully
this helps, but I have not tested it yet.

In the future, we should eventually move to zod v4 anyway, so it seems
like a low risk for a potential benefit.

https://v4.zod.dev/v4