fix: don't try to show neovim :messages as they cannot be seen (#499)
After https://github.com/mikavilpas/tui-sandbox/pull/478, the output is
not shown by default. It makes no sense to try to show it in the
afterEach hook, as it will not be visible to the user.

I think it also sometimes blocked the test run despite my attempt to use
a timeout. Better to get rid of it.