feat: can start neovim with a different NVIM_APPNAME
**Issue:**

It's not possible to start a Neovim test session with different sets of
plugins and LSP servers installed. Some plugins might look for the
presence of a plugin, and enable or disable features based on that.

However, if a plugin is used in a different test, it will be available
in the environment, and the test will exhibit different behavior than
what would have happened if the plugin was not installed.

**Solution:**

Now it's possible to start Neovim with a different `init.lua`
configuration file by setting the `NVIM_APPNAME` environment variable.

```ts
// loads the file test-environment/.config/nvim_alt/init.lua
cy.startNeovim({ NVIM_APPNAME: "nvim_alt" })

// these load the file test-environment/.config/nvim/init.lua
cy.startNeovim({ NVIM_APPNAME: "nvim" })
cy.startNeovim({}) // default NVIM_APPNAME is "nvim"
```

https://neovim.io/doc/user/starting.html#_nvim_appname