fix: rendering emoji as truncated images
**Issue:**

In neovim, displaying an emoji such as 🎉 is displayed as a double-width
character with part of the second half missing, resulting in a truncated
image.

Neovim is able to display emoji perfectly fine outside of xterm.js.

**Solution:**

Add the `@xterm/addon-unicode11` addon to xterm.js, which fixes the
emoji rendering issue by enabling Unicode 11 support.