feat: `tui run` command for running tests
**Issue:**

It's complicated to set up an environment for running tests with
tui-sandbox.

I want to start reducing the amount of time and knowledge that users
need to have to use tui-sandbox.

**Solution:**

As a first step, add a `tui run` command that

- starts the tui-sandbox server
- waits for it to be ready
- runs the tests using Cypress

Now users can depend on this simple command to run tests, instead of
having to set up and maintain the different components of the
environment themselves.