feat: `textIsVisibleWithColor` matches can be hovered in cypress log (#557)
**Issue:**

When matching text in the terminal with `textIsVisibleWithColor` or
`textIsVisibleWithBackgroundColor`, the matches are not highlighted in
the Cypress log, making it difficult to visually verify which pieces of
text were matched.

**Solution:**

Allow the matches to be hovered in the Cypress log. Cypress will then
highlight the matches on the page.