fix: sometimes processing keypresses out of order (#881)
**Issue:**

When sending a long sequence of keypresses quickly to Neovim, they may
sometimes be processed out of order. This probably happens because the
order of the async calls to sendStdin is not guaranteed.

**Solution:**

Introduce a BatchedAsyncQueue that batches keypress events and ensures
they are sent to Neovim in the order they were received.

This guarantees the order of keypresses to match the order of user
input. It can be used quite performantly when using a low delay:

```ts
// everything is sent in a single batch when using { delay: 0 }
//
// If a larger delay is used, multiple batches are likely sent, and
// performance may degrade.
cy.typeIntoTerminal(`i${longText.repeat(20)}{esc}`, { delay: 0 })
```

- bump target and lib to ES2024 in tsconfig.json to get access to
  `Promise.withResolvers()`