refactor: use static create method for neovim & terminal clients
**Issue:**

Both `NeovimTerminalClient` and `TerminalTerminalClient` have a
constructor that needs to initialize asynchronous resources.

Because constructors cannot be async, they currently have a
`ready: Promise<void>` variable that needs to be awaited when using any
of the client's methods.

**Solution:**

In both cases, use a static `create` method that handles the
asynchronous initialization. The constructor now only stores the
resources that are ready to be used.

This makes the code easier to understand.